% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR_code.R
\name{odataR_get_cat}
\alias{odataR_get_cat}
\title{Get catalog information}
\usage{
odataR_get_cat(root = NULL, type = "Tables", query = NULL, debug = F)
}
\arguments{
\item{root}{Character string with root of structure. Default NULL i.e. determined by the last use of \code{odataR_set_root} and if never used \code{'https://opendata.cbs.nl'}}

\item{type}{Type of catalog information. This type can be one of:
\itemize{
 \item \code{NULL} : a \code{tibble} with the various types in this catalog \cr
 \item \code{'Tables'} or empty : a \code{tibble} with the Tables in this catalog \cr
 \item \code{'Featured'} : a \code{tibble} with the Features in this catalog \cr
 \item \code{'Table_Featured'} : a \code{tibble} with the Table_Featured in this catalog \cr
 \item \code{'Themes'} : a \code{tibble} with the Themes in this catalog \cr
 \item \code{'Tables_Themes'} : a \code{tibble} with the Tables_Themes in this catalog \cr
}}

\item{query}{OData query to restrict data returned from the catalog}

\item{debug}{Boolean indicating if (for debugging) the generated OData queries are to be printed (default FALSE)}
}
\value{
A \code{tibble} (\code{data_frame}) when successful otherwise an empty \code{tibble} or a message
}
\description{
This function can be used to get information from an OData catalog.
}
\section{Remarks}{

See  \url{http://www.odata.org/documentation/odata-version-3-0/odata-version-3-0-core-protocol/} or \url{http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html} (depending on the version of the OData protocol that is used) for details about the query possibilities.

\code{$format=json} is forced by the code so do not specify \code{$format=atom} because this will not work
}

\examples{
\dontrun{
# all types in catalog:
df = odataR_get_cat(type=NULL)
# all Tables in catalog:
df = odataR_get_cat()
# all tables that have a Title starting with 'Vacature':
df = odataR_get_cat(query="?$filter=startswith(Title,'Vacature')")
# all tables that have a Title containing 'vacature' regardless of case:
df = odataR_get_cat(query="?$filter=substringof('vacature',tolower(Title))")
# all active tables that have a Title containing 'vacature' regardless of case:
df = odataR_get_cat(query="?$filter=substringof('vacature',tolower(Title)) and Frequency ne 'Stopgezet' ")
# all Featured entries in catalog:
df = odataR_get_cat(type='Featured')
# all Featured tables (?) in catalog:
df = odataR_get_cat(type='Table_Featured')
# all Themes in catalog:
df = odataR_get_cat(type='Themes')
# all Tables_Themes (?) in catalog:
df = odataR_get_cat(type='Tables_Themes')
}
}
