% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR_code.R
\name{odataR_get_meta}
\alias{odataR_get_meta}
\title{Get meta data about table}
\usage{
odataR_get_meta(
  root = NULL,
  table_id = NULL,
  metatype = NULL,
  query = NULL,
  debug = FALSE
)
}
\arguments{
\item{root}{Character string with root of structure. Default NULL i.e. determined by the last use of \code{odataR_set_root} and if never used \code{'https://opendata.cbs.nl'}}

\item{table_id}{Identification of table}

\item{metatype}{indicates the type of metadata that will be retrieved. Possible values are:
\itemize{
 \item \code{NULL} or empty : a \code{tibble} with the urls of the subtables will be retrieved \cr
 \item \code{'TableInfos'} : a \code{tibble} with one row is retrieved with (background) information about the table \cr
 \item \code{'CategoryGroups'} : a \code{tibble} with one row is retrieved with information about the category tables for this table \cr
 \item \code{'DataProperties'} : a \code{tibble} is retrieved with information about the columns of the table \cr
 \item a dimension name : a \code{tibble} is retrieved with the code information of the dimension (e.g.  \code{Gemeenten} in the examples below). \cr
}}

\item{query}{OData query to restrict data returned from the meta table (do not use in combination with \code{metatype=NULL})}

\item{debug}{Boolean indicating if (for debugging) the generated OData queries are to be printed (default FALSE)}
}
\value{
A \code{tibble} (\code{data_frame}) when successful otherwise an empty \code{tibble} or a message
}
\description{
This function can be used to get information about an OData table.
}
\section{Remarks}{

See  \url{http://www.odata.org/documentation/odata-version-3-0/odata-version-3-0-core-protocol/} or \url{http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html} (depending on the version of the OData protocol that is used) for details about the query possibilities.

\code{$format=json} is forced by the code so do not specify \code{$format=atom} because this will not work
}

\examples{
\dontrun{
odataR_set_root('http://dataderden.cbs.nl') 
# NB statement in previous line is necessary
#  otherwise the examples below will not work!
df = odataR_get_meta(table_id='45042NED')
df = odataR_get_meta(table_id='45042NED',metatype='TableInfos')
df = odataR_get_meta(table_id='45042NED',metatype='CategoryGroups')
df = odataR_get_meta(table_id='45042NED',metatype='DataProperties')
df = odataR_get_meta(table_id='45042NED',metatype='Gemeenten')
df = odataR_get_meta(table_id='45042NED',metatype='Verslagsoort')
df = odataR_get_meta(table_id='45042NED',metatype='Gemeenten',
       query="?$filter=substringof('amstelveen',tolower(Title))",
       debug=T)
}
}
