% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{odataR_get_root_catalog}
\alias{odataR_get_root_catalog}
\title{Gets the root for OData catalog structure}
\usage{
odataR_get_root_catalog(type = "Tables")
}
\description{
Determines from which structure the catalog information will be extracted. Derived from the common OData data and catalog root: \code{paste0(odataR_get_root(), '/ODataCatalog/Tables')}.
}
\details{
Types of catalog entries available are \code{Featured}, \code{Table_Featured}, \code{Tables}, \code{Themes}, \code{Tables_Themes}
}
\seealso{
\code{\link{odataR_get_root}} and \code{\link{odataR_get_root_data}}
}
