% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{odataR_get_root_data}
\alias{odataR_get_root_data}
\title{Gets the root for OData data structure}
\usage{
odataR_get_root_data()
}
\description{
Determines from which structure the data will be extracted. Derived from the common OData data and catalog root: \code{paste0(odataR_get_root(), '/ODataFeed/OData')}
}
\seealso{
\code{\link{odataR_get_root}} and \code{\link{odataR_get_root_catalog}}
}
