% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR_code.R
\name{odataR_get_table}
\alias{odataR_get_table}
\title{Get table data}
\usage{
odataR_get_table(
  root = NULL,
  table_id = NULL,
  query = NULL,
  typed = T,
  decode = T,
  keepcode = c(),
  debug = F
)
}
\arguments{
\item{root}{Character string with root of structure. Default NULL i.e. determined by the last use of \code{odataR_set_root} and if never used \code{'https://opendata.cbs.nl'}}

\item{table_id}{Character string with identification of table}

\item{query}{OData query to restrict data returned from structure}

\item{typed}{Boolean indicating 'TypedDataSet' when T or 'UntypedDataSet' when F}

\item{decode}{Boolean indicating if fields are to be decoded (default T)}

\item{keepcode}{Character string with dimension(s) for which the coded values are kept}

\item{debug}{Boolean indicating if (for debugging) the generated OData queries are to be printed (default FALSE)}
}
\value{
A \code{tibble} (\code{data_frame}) when successful otherwise an empty \code{tibble} or a message
}
\description{
This is the main function of this package. It retrieves a data table from the data base and all dimensions are decoded (if requested). An (optional) query is executed in the database
}
\section{Remarks}{

See  \url{http://www.odata.org/documentation/odata-version-3-0/odata-version-3-0-core-protocol/} or \url{http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html} (depending on the version of the OData protocol that is used) for details about the query possibilities.

\code{$format=json} is forced by the code so do not specify \code{$format=atom} because this will not work
}

\examples{
\dontrun{
df      = odataR_get_table(table_id="82935NED")
df      = odataR_get_table(table_id="82935NED",keepcode = "RegioS")
df      = odataR_get_table(table_id="82935NED",keepcode = c("RegioS","Perioden"))
df      = odataR_get_table(table_id="82935NED",
  query  = paste0("?$filter=startswith(RegioS,'NL01')",
                  "&$select=RegioS,Perioden,TotaleInvesteringen_1",
                  "&$skip=2&$top=3"),
  debug=T)
}
}
